refactor: extract joinPath, dirname, and affectedDirsForpath to pathUtils
- the three path utilities were unrelated to React state and was just
makng the hook complex so extracting them reduces the hook by ~36 lines
and makes the watcher refresh loop cleaner (will move the watcher
lifecycle out after this too)
- i also created test for pathUtils locally but won't commit it since we
don't have testing in the frontend yet.
